+++ tests/lib/outputs/csv/csv_test.py
@@ -4,60 +4,59 @@ from unittest.mock import Mock
 
 import pytest
 
-from prowler.lib.outputs.csv.csv import write_csv
-from prowler.lib.outputs.output import CSV, Finding, Output, Severity, Status
-
-findings = [
-    Finding(
-        auth_method="OAuth",
-        timestamp=datetime.now(),
-        account_uid="12345",
-        account_name="Example Account",
-        account_email="example@example.com",
-        account_organization_uid="org-123",
-        account_organization_name="Example Org",
-        account_tags=["tag1", "tag2"],
-        finding_uid="finding-123",
-        provider="AWS",
-        check_id="check-123",
-        check_title="Example Check",
-        check_type="Security",
-        status=Status("FAIL"),
-        status_extended="Extended status",
-        muted=False,
-        service_name="Example Service",
-        subservice_name="Example Subservice",
-        severity=Severity("critical"),
-        resource_type="Instance",
-        resource_uid="resource-123",
-        resource_name="Example Resource",
-        resource_details="Detailed information about the resource",
-        resource_tags="tag1,tag2",
-        partition="aws",
-        region="us-west-1",
-        description="Description of the finding",
-        risk="High",
-        related_url="http://example.com",
-        remediation_recommendation_text="Recommendation text",
-        remediation_recommendation_url="http://example.com/remediation",
-        remediation_code_nativeiac="native-iac-code",
-        remediation_code_terraform="terraform-code",
-        remediation_code_cli="cli-code",
-        remediation_code_other="other-code",
-        compliance={"compliance_key": "compliance_value"},
-        categories="category1,category2",
-        depends_on="dependency",
-        related_to="related finding",
-        notes="Notes about the finding",
-        prowler_version="1.0",
-    )
-]
+from prowler.lib.outputs.csv.csv import CSV, write_csv
+from prowler.lib.outputs.finding import Severity, Status
+from prowler.lib.outputs.output import Finding, Output
 
 
 class TestCSV:
+    def generate_finding(self):
+        return Finding(
+            auth_method="OAuth",
+            timestamp=datetime.now(),
+            account_uid="12345",
+            account_name="Example Account",
+            account_email="example@example.com",
+            account_organization_uid="org-123",
+            account_organization_name="Example Org",
+            account_tags=["tag1", "tag2"],
+            finding_uid="finding-123",
+            provider="AWS",
+            check_id="check-123",
+            check_title="Example Check",
+            check_type="Security",
+            status=Status("FAIL"),
+            status_extended="Extended status",
+            muted=False,
+            service_name="Example Service",
+            subservice_name="Example Subservice",
+            severity=Severity("critical"),
+            resource_type="Instance",
+            resource_uid="resource-123",
+            resource_name="Example Resource",
+            resource_details="Detailed information about the resource",
+            resource_tags="tag1,tag2",
+            partition="aws",
+            region="us-west-1",
+            description="Description of the finding",
+            risk="High",
+            related_url="http://example.com",
+            remediation_recommendation_text="Recommendation text",
+            remediation_recommendation_url="http://example.com/remediation",
+            remediation_code_nativeiac="native-iac-code",
+            remediation_code_terraform="terraform-code",
+            remediation_code_cli="cli-code",
+            remediation_code_other="other-code",
+            compliance={"compliance_key": "compliance_value"},
+            categories="category1,category2",
+            depends_on="dependency",
+            related_to="related finding",
+            notes="Notes about the finding",
+            prowler_version="1.0",
+        )
 
     def test_output_transform(self):
-        print(findings)
+        findings = [self.generate_finding()]
         output = CSV(findings)
         output_data = output.data[0]
         assert isinstance(output_data, dict)
@@ -115,10 +114,10 @@ class TestCSV:
         assert output_data["prowler_version"] == "1.0"
 
     def test_csv_write_to_file(self):
-        print(findings.__class__.__name__)
-        output = CSV(findings)
         mock_file = StringIO()
-        output.write_to_file(mock_file)
+        findings = [self.generate_finding()]
+        output = CSV(findings, create_file_descriptor=True, file_path=mock_file)
+        output.batch_write_findings_to_file()
 
         mock_file.seek(0)
         content = mock_file.read()
@@ -173,9 +172,6 @@ class TestCSV:
         assert dummy_output.transform(finding_example) is None
         with pytest.raises(NotImplementedError):
             dummy_output.write_to_file(Mock())
-
-
-class TestWriteCSV:
 
     def test_write_csv_with_dict(self):
         headers = ["provider", "account", "check_id"]
